Displaying first Quad

+++ Aether Renderer/Engine/Mesh.cpp
@@ -1,6 +1,6 @@
 #include "Mesh.h"
 
-Mesh::Mesh(std::vector<float> vertices, std::vector<float> indices)
+Mesh::Mesh(std::vector<float> vertices, std::vector<unsigned int> indices)
 {
 	Vertices = vertices;
 	Indices = indices;
